added the GUI
doesnt fully work but i got the auto, manual, and exit button to all show up.
there is a background overlapping the colored background i had so im looking for a way to remove that. the up and down arrows i added are no functioning properly so i turned them off in GUIMenu.